[ConstraintSystem] Add the notion of disjunction partitioning.
Update DisjunctionChoiceProducer so that it creates a sorted ordering
of the choices in the disjunction as well as a partitioning of those
choices.

If we hit the end of a partition and have found a solution, we'll stop
attempting choices from this disjunction.

For now we'll default to keeping the same order that the disjunctions
initially have, and will put them all into the same partition. A
future commit will implement logic to implement more interesting
partitions of the choices.