[Mangling] Mangle protocol names in assoc-type-paths.
The mangling of associated type paths was only adding the names of
associated types, and not their enclosing protocols. This led to mangling
collisions that could lead to corrupted metadata. In the standard
library, for example, the generic requirements for the 
Unicode _ParsingIterator in the standard library ended up encoding an 
access to Sequence.Element rather than IteratorProtocol.Element due
to the mangling conflict.

Part of SR-7553 / rdar://problem/39769906.