[Sema] Don't unwrap CoerceExpr from literal coercion
When converting an initialization to a coercion like:

```swift
_ = Float(90)
```

the outer coercion is given the startLoc and endLoc of the CallExpr in
source.

However, once we unwrap it and pull the implicit CallExpr to
`init(_builtinIntegerLiteral:)`, we lose the original source range.

Instead, leave the outer CoerceExpr, as it's ignored in SILGen anyway.