Remove an overly fragile test for armv7k
We're not actually sure what this test is trying to validate, but it's quite fragile and probably not the best way to test it. Remove it rather than continuing to blindly update it.

Also minor improvements for the generic converts on 32b platforms; there's still more to be done with these, but this keeps us on the HW path for the common case.