[Runtime] Make Swift errors bridged to NSError and cloned round-trappable.
When a Swift error is bridged to NSError, it maintains enough information
to round-trip back to Swift. However, if the NSError is then cloned—e.g.,
a new NSError is created with the same domain/code/user-info dictionary,
the identity of the Swift type is lost.

When we initially form the user-info dictionary for a Swift-defined
error that is not explicitly bridgeable back from an NSError instance,
add a new key to the user-info dictionary containing a wrapped copy of the
original error. When trying to bridge from NSError, look for that key in
the user-info dictionary and cast to its value.

Fixes rdar://problem/29882046.