In eager mode, added support to running a Tensor<Float>(x) statement.
This involves refactoring the scalar promotion code, and also fixing a bug in
const eval over an alloc_stack inst. The details are as follows.

Previously, `Tensor<Float>(x)` compiles to a tfc.scalarToTensor graph_op. It is a
pseudo graph op in that it is not backed by any TF ops. Instead, the partition
pass (SILCloner) lowers tfc.scalarToTensor to scalar-promoted graph nodes. As
such, in eager mode, this graph_op has no runtime implementation that IRGen can
dispatch to.

In this patch, we changed the above Tensor initializer on scalar to generate a
call to a new function __tf_tensor_from_scalar(scalar), which has a runtime impl. In
the deabstraction pass, it is either transformed to a "Const" graph node, if `x`
is a compile-time constant, or otherwise a `tfc.scalarToTensor` graph_op, so
that the existing scalar promotion infra in the partition pass continues to work.

The following example shows the importance of representing this scalar-to-tensor
transformation as a graph_op (as opposed to, say an apply inst of some function)
in the current scalar promotion infrastructure.

```swift
let x_scalar = x_tensor.mean()
let y_scalar = y_tensor.mean()
let z_scalar = x_scalar + y_scalar
let z_tensor = Tensor(z_scalar)
```

The scalar promotion code starts at the tfc.scalarToTensor graph_op inst
associated with `Tensor(z_scalar)`, and transitively promotes the operands and
their producer insts to the graph. In this case, the scalar add is promoted to
graph, and the host-side scalars `x_scalar` and `y_scalar` are eliminated, since
they are produced by _getScalarOrDie() calls. As such, the
tfc.scalarToTensor graph_op and the _getScalarOrDie() call that feeds the
graph_op "cancel out" on each other, such that all intermediate computation is
promoted into the graph.

While it is possible to refactor the scalar promotion code to work on an apply
inst over __tf_tensor_from_scalar() instead of a tfc.scalarToTensor graph_op,
doing so would be inconvenient (and thus add code base complexity). In
particular, __tf_tensor_from_scalar() is generic over the Scalar data
type (any type that conforms to AccelerableByTensorFlow), and therefore will be
taking the input `scalar` indirectly. See the `@in_guaranteed` param in the
example SIL below:

```
  %702 = function_ref @__tf_tensor_from_scalar : $@convention(thin) <Ä_0_0 where Ä_0_0 : AccelerableByTensorFlow> (@in_guaranteed Ä_0_0) -> @owned TensorHandle<Ä_0_0> // user: %703
  %703 = apply %702<Int32>(%700) : $@convention(thin) <Ä_0_0 where Ä_0_0 : AccelerableByTensorFlow> (@in_guaranteed Ä_0_0) -> @owned TensorHandle<Ä_0_0> // users: %710, %707, %738, %708, %706, %705
```

To make scalar promotion work on the __tf_tensor_from_scalar() call instead of
the tfc.scalarToTensor graph_op, the changes include:
1. Add apply insts over __tf_tensor_from_scalar() to tensor_ops
2. Find the producer of the (indirect) input scalar operand of
__tf_tensor_from_scalar(), and recursively promote the producer of that scalar.

Since the param is indirect, step `#2` involves some SIL code traversal
logic (similar to the const evaluator) to find the producer inst of the
scalar. In the SIL example below, we start at `%700` used by the
__tf_tensor_from_scalar() call above, and traverse back to `%58` (and possibly
further). This would add non-trivial complexity.

```
  %694 = begin_access [read] [static] %58 : $*Int32 // users: %695, %696
  %695 = load %694 : $*Int32                      // user: %698
  end_access %694 : $*Int32                       // id: %696
  %697 = alloc_stack $Int32                       // users: %711, %709, %699, %698
  store %695 to %697 : $*Int32                    // id: %698
  %699 = load %697 : $*Int32                      // user: %701
  %700 = alloc_stack $Int32                       // users: %704, %703, %701
  store %699 to %700 : $*Int32                    // id: %701
```

To summarize, we attempted to eliminate tfc.scalarToTensor, but concluded that
did not reduce code complexity. The lifetime of tfc.scalarToTensor insts is well
constrained (and thus its associated code complexity is contained) -- it only
occurs in graph mode, where it's generated in deabstraction, and consumed in
partition. It does not appear in graph lowering.

Earlier we also tried having `Tensor<Float>(x)` continue to generate
tfc.scalarToTensor, and transform it to a __tf_tensor_from_scalar() call in
deabstraction when needed. We find it significantly more challenging to
synthesize a call to __tf_tensor_from_scalar() than to synthesize a graph_op
inst.

The main complexity comes from protocol conformance
requirements. _tf_tensor_from_scalar() is generic over Scalar type, and since
Scalar conforms to AccelerableByTensorFlow, we cannot create a SubstitutionMap
via tf::getSingleSubstitutionMapForElementType() (as we did in
createHostReceive() any various other parts in the partition pass). Instead, we
have to write code like (thanks for pointers from rxwei@):

```
auto &ctx = B.getASTContext();
auto *tfModule = ctx.getLoadedModule(ctx.Id_TensorFlow);
auto proto =
    ctx.getProtocol(KnownProtocolKind::AccelerableByTensorFlow);
SmallVector<ProtocolConformance *, 1> conformances;
auto nominal = scalarValueTy->getAnyNominal();
auto lookup =
    nominal->lookupConformance(tfModule, proto, conformances);
if (!lookup) {
  llvm_unreachable("No conformance to AccelerableByTensorFlow found.");
}
assert(conformances.size() == 1 &&
       "Found multiple conformance candidates.");
auto *conf =
    ctx.getConformance(scalarValueTy, proto, nominal->getLoc(),
                       nominal, ProtocolConformanceState::Complete);
ProtocolConformanceRef confRef(conf);
// Get substitutions.
auto subMap = SubstitutionMap::getProtocolSubstitutions(
    proto, scalarValueTy, confRef);
```

Even with the above, we got linker errors complaining symbols like
`SSf10TensorFlow013AccelerableByaB0sWa` cannot be found.

 swift-demangle tells us:
```
$ swift-demangle   SSf10TensorFlow013AccelerableByaB0sWa
$SSf10TensorFlow013AccelerableByaB0sWa ---> protocol witness table accessor for Swift.Float : TensorFlow.AccelerableByTensorFlow in Swift
```

A possible fix is to call additional code like the following, but we have not
tried that.
```c++
  auto Ret = lookUpWitnessTable(C);
```

Overall there is not much existing compiler code to follow that generates calls
to functions with protocol conformance requirements. The code in SILGen in that
regards seems tightly packaged such that we haven't been able to identify (and
refactor) the relevant parts for us to use in other compiler phases/passes.

This PR exposed an existing bug in const eval, which was written up at
https://github.com/apple/swift/pull/19661. The last test in
"DatasetAPITests.testAllBackends("DoubleValueDatasetIteration")" would fail, and
in non-deterministic ways (release and debug versions of the compiler behave
differently).

We created a short-term fix in this patch, and it will be relaced by a more
thorough fix from marcrasi@ shortly.