[AST] Tail-allocate space for the files in a ModuleDecl
...removing the need for a destructor (and also shrinking the type a
little). This does mean you need to know how many files are in a
module up front, but it turns out we /do/ always know that, and moreover
that it's usually "1".

There is a place where we don't know this number exactly, and it's
code completion: in some cases, the code completion buffer isn't
included in the normal inputs. I'll clean that up next.