Allow synthesis of gradient functions using a subset of adjoint indices.
Each gradient instruction now creates a canonical (and potentially 
non-canonical) gradient function with the requested set of indices. This 
function will call the primal and adjoint with the minimal index 
superset if one exists, or synthesize a primal/adjoint with exactly the 
requested set if not. Also adds a test for this behavior.